@@ -1,6 +1,6 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 import { routes } from './router';
 import { AppComponent } from './app.component';
@@ -10,15 +10,9 @@ import { RouterModule } from '@angular/router';
 import { externalModules } from './build-specific';
 import { declarations } from './core';
 
-@NgModule({
-  declarations: [AppComponent, AboutComponent, declarations],
-  imports: [
-    BrowserModule,
-    HttpClientModule,
-    RouterModule.forRoot(routes, {}),
-    AppStoreModule,
-    externalModules
-  ],
-  bootstrap: [AppComponent]
-})
+@NgModule({ declarations: [AppComponent, AboutComponent, declarations],
+    bootstrap: [AppComponent], imports: [BrowserModule,
+        RouterModule.forRoot(routes, {}),
+        AppStoreModule,
+        externalModules], providers: [provideHttpClient(withInterceptorsFromDi())] })
 export class AppModule {}
